5.16. Модуляция
-изменил тип подключаемого скрипта
-добавил 3 файла(photo-data.js, static-data.js, util.js)
-разбил содержимое main.js на модули
-изменил название функции createDescriptionPhoto, на createDataPhoto(создаваемый объект больше содержит данные, чем описание).
