+++ js/static-data.js
@@ -1,0 +1,10 @@
+const MIN = 2;
+const MAX = 24;
+const STRING = 'Проверяемая_строка';
+const MAX_STRING_LENGTH = 140;
+const MAX_COUNT_OBJECTS = 25;
+const MIN_RANDOM = 15;
+const MAX_RANDOM = 200;
+const DESCRIPTIONS = ['фото с отпуска','закат','цветы','пляж'];
+
+export {MIN,MAX,STRING,MAX_STRING_LENGTH,MAX_COUNT_OBJECTS,MIN_RANDOM,MAX_RANDOM,DESCRIPTIONS};
